pyproject.toml
[project]
name = "flightshub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg==0.31.0",
    "fastapi>=0.124.4",
    "fastapi-pagination>=0.15.4",
    "httpx>=0.28.1",
    "psycopg2>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-extra-types>=2.11.0",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sqlalchemy==2.0.45",
    "sqlmodel>=0.0.27",
    "strawberry-graphql[fastapi]>=0.289.2",
    "testcontainers[postgresql]>=4.13.3",
    "tzdata>=2025.3",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "docker>=7.1.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
